Fix Journal cursor deadlock
The Journal daemon plugin keeps a Mutex around the cursor of where it last read
from the journal.

However, when updating the session we attempt to take a lock on the
cursor before the first one has been dropped.

Wrap the acquistion of the cursor and clone the contents, so it's
expclitly dropped.

In addition, read all the journal response into text and then convert to
messages instead of streaming, as this is what we end up doing anyway.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>