Only destroy matching sessions
If the iml-agent gets a `SESSION_DESTROY` message, it will destroy the
session with a matching plugin name, even if the session id does not
match.

This could lead to the wrong session being destroyed.

Check if the currently held session id matches the one to be destroyed.
If so, destroy it, if not it's a noop.

In addition, Remove the `SessionInfo` struct and replace `Seq` with
`AtomicU64`. This allows us to drop a Mutex, as `AtomicU64` is lockless
and threadsafe.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>