Ensure empty snapshots do not producce parse error
If `snapshot_list` is called when there are no snapshots, no output is
produced.

If this is fed into `parse_snapshot_list` it will produce an error like:

```
Parse error at 0
Unexpected `end of input`
Expected `:` or `whitespaces`
```

The easiest way to head this off is to make sure the input is non-empty
and return an empty `Vec` if it is.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>