Use IML cli inside iml-timer
The iml-timer service currently uses it's own internal CLI for stratagem
scans instead of reusing the iml cli.

If we follow this pattern, it will mean we need to create a separate
binary for each service we want to run as a timer inside of docker.

Instead, we should embed the iml cli inside the container and call it
the same way we do outside the container.

This will allow us to write a single cli command for running snapshot
commands instead of having to maintain two commands.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>