Fix Stat deletion
The current targets table is intermittently losing it's filesystems for
some targets.

This appears to be because we delete stat info before inserting the next
batch. If anyone queries during this gap, they will get empty data even
though we have the points, they just haven't been inserted yet.

In addition, we are clearing out all stats except a single point for any target colocated on the MGS node.

This patch waits until after insertion to do a deletion, and scopes the
delete so it should only effect the correct stat.

Fixes #2258.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>