GH-2268 Update the limit param in graphql api's (#2269)
* GH-2268 Update the limit param in graphql api's

Fixes #2268.

Currently, the limit param allows you to specify a limit when retrieving items back from an API call. This is a numeric value that can be passed in but in some cases, we may want all results. To address this, postress allows us to pass in ALL instead of a numeric value, which is effectively the same as not passing a limit at all. Update all of the graphql API endpoints such that if None is passed in it will pass ALL to the LIMIT clause in the query. This will allow us to get all entries without modifying the existing api.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Small update

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Fix formatting

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Update limit comment to read 'All' instead of '20'

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Update limit description

Signed-off-by: johnsonw <wjohnson@whamcloud.com>