EX-1932 handle softlinks for llapi queries
When constructing an llapi instance from a random path
liblustreapi first resolves that path fully, including
symbolic links.  This causes a failure when trying to
get an llapi instance from a symbolic link that lives on
a lustre filesystem (ex /mnt/lustre/symbolic_link)

If the path passed into search_fsname is a link, use
the parent dir instead.

Signed-off-by: Ben Evans <beevans@whamcloud.com>