Implement full stratagem scans in single mutation
Running a scan currently consists of at least three different areas on
the manager:

1. The python API which performs validation and resource -> db id
resolution
2. The job_scheduler, which assigns tasks and builds out the command
plan.
3. The RunStratagem step, which will build the actual stratagem config and ship it
off to each MDT.

Each area only owns part of the whole solution (creating /
destroying tasks and shipping a config off to an MDT). It is overly
complex to split responsibility in this way, and also brittle for future
changes.

Instead, we can add a single mutation that takes in the fsname to scan
+ tasks and rule groups and does all the work necessary to
generate the command. This cuts the job_scheduler out of the picture and
the step to run the config becomes very simple as it no longer is
responsible for building the config.

- Create a graphql mutation to run stratagem scans given a config and
some tasks.
- Extract stratagem api code to a separate module.
- Implement a simple step to scan a MDT.
- Refactor GUI to remove reliance on certain profiles `USE_STRATAGEM=true` is now the only requirement.
- Refactor CLI to use graphql
- Refactor GUI to use graphql
- Refactor integration testing to share more of the path

Signed-off-by: Joe Grund <jgrund@whamcloud.io>