Stop client changes when adding / removing servers
Due to the use of reverse_dependencies, existing clients want to be unmounted
when adding a new server. This should not need to happen, but in
particular it should not happen when there is no filesystem detected.
Stop the reverse dependencies from propagating when there is no fs
detected yet.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>