Update action-client
Update action-client once again to optionally take a uuid that can be
used to bookkeep the request in flight.

In addition, remove `invoke_rust_agent_cancelable` as it adds an extra
layer of indirection that is more useful in tools calling the client.

Replace it with a `cancel_request` fn that will take a `Uuid` ref and
attempt to cancel a remote rpc matching the uuid.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>