Duplicate target insert during attempted target mount
When targets are mounted quickly in succession (i.e after a failed mount), we can get data
from the agents showing the target mounted on both hosts.
When this occurs we will try to insert the target twice into the db.

Instead we should filter out any items that show as mounted but are unknown to the current host.

Fixes #2380

Signed-off-by: Joe Grund <jgrund@whamcloud.io>