Fix Stale Influx Entries (#2373)
* Fix Stale Influx Entries

There is currently a case in which a stale entries are being left in the
influx db. This takes place when a filesystem is mounted, data is pulled
into influx, and then another filesystem is mounted, and both
filesystems use the same mgs. For example:

time n:
mgs_fs: "fs"

time n+1:
mgs_fs: "fs2"

time n+2:
mgs_fs: "fs,fs2"

Currently, only items matching "fs,fs2" will be removed, since that's
the current input. This results in entries in the target table showing
that the filesystems are: "fs, fs2, fs,fs2".

This patch ensures that all subsets of the input filesystems are deleted
and all but the last of the current subset is deleted.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* CR updates

Signed-off-by: johnsonw <wjohnson@whamcloud.com>