Fix client mount generation (#2450)
Latest version of EXAScaler appears to have a number of new resource
constraints that weren't in the previous version.

This in turn has exposed a bug where we are not filtering by banned resource properly
and are picking up all target bans, not just the MGS one.

Instead, we should match the resource mount_point with the target
mount_path to only include the relevant resource.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>