Implement static schema and precise validation (#213)
* Implement static schema and precise validation

The state-schema is currently a .yml file that is both parsed and loaded
at the start of runtime.

After thinking about this a bit, there is no reason why the state-schema
needs to be implemented in this way. The state-schema is only
meant for developer usage.

Instead, we can define the state-schema in rust, which will allow us to
catch any errors at compile time instead of runtime / in unit tests.

In addition, we can use rust for validation of input documents as well,
rather than using jsonschema. This should allow for a better user
experience.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>
Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Update StateSchema structure

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Make TransitionGraph types ord just so tests are stable.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* - Make "tags" optional in client_mount Create input
- Add MgtType enum to filesystem to ensure only the appropriate value
can be passed in
- Add input for "unmount" action in mgt
- Add remaining action match arms
- Add tests for each component
- Add some validation tests. There are still a few more I believe I can
add in a follow up patch.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Remove unused code and format

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* - Testing updates

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - fix clippy errors

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Remove unused schemas

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

Co-authored-by: johnsonw <wjohnson@whamcloud.com>