EX-2480 - API - Create endpoint to receive input documents (#196)
* EX-2480 - API - Create endpoint to receive input documents

The state machine must have a way to receive input documents. These
documents can be submitted to the state machine by using a graphql
endpoint. When the call is received, two things should happen:

1. Record the submitted input document in jsonb format into the
"input_document" table if validation passes
2. Submit the input document to the state machine for processing

After the document is submitted, the state machine should return a
Command object. This part will be handled in a separate ticket.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>
Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* update sqlx-data

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

Co-authored-by: Joe Grund <jgrund@whamcloud.io>