Implement command plan persistence
Fixes [EX-2786](https://jira.whamcloud.com/browse/EX-2786)

Implement persistence of running command plans.
This is implemented as a main db receiver and a series of scoped writers
so many updates to the plan can be queued in parallel.

In addition, implement return of a serialized command plan and rendering
in the CLI.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>