Fix emf-ssh spawn tests (#247)
* - The spawn integration tests were failing due to the tests starting a
server on the same port in parallel. I've updated the tests to run on
separate ports so that they can still run in parallel but not overlap
each other.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - The stderr message is different depending on which operating system
the test is being run on. We just need to make sure that the buffer
contains "stderr: ls:" and that should satisfy the test regardless of
the operating system.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - The stderr test returns an exit code of 1 on OS X and an exit code of
2 on linux. Account for both.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>